[Laravel v8.0] Completely refactored Eloquent Factories.
